pyproject.toml
[project]
name = "lgc-nb-additions"
dynamic = ["version"]
authors = [{ name = "LgCookie", email = "lgc2333@126.com" }]
dependencies = [
    "cookit[loguru]>=0.11.1",
    "nonebot-plugin-alconna>=0.57.6",
    "nonebot2>=2.4.1",
    "pydantic>=2.11.4",
    "python-debouncer>=0.1.5",
]
requires-python = ">=3.12,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/lgc2333/nb-additions"

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "lgc_nb_additions/__init__.py"
